+++ main_test.go
@@ -22,10 +22,10 @@ package main
 
 import (
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/urfave/cli"
-	"time"
 )
 
 func Test_createConfig_args_narg1(t *testing.T) {
@@ -90,6 +90,7 @@ func Test_createConfig_flags_empty(t *testing.T) {
 
 	// then
 	assert.NoError(t, err)
+	assert.Nil(t, config.creds)
 	assert.False(t, config.noFail)
 }
 
@@ -97,6 +98,7 @@ func Test_createConfig_flags(t *testing.T) {
 	// given
 	args := cli.Args{"foo"}
 	flags := &appFlags{
+		tls:     true,
 		noFail:  true,
 		timeout: time.Minute,
 	}
@@ -106,6 +108,70 @@ func Test_createConfig_flags(t *testing.T) {
 
 	// then
 	assert.NoError(t, err)
+	assert.NotNil(t, config.creds)
 	assert.Equal(t, time.Minute, config.timeout)
 	assert.True(t, config.noFail)
+}
+
+func Test_parseCredentials_tls(t *testing.T) {
+	// given
+	dataset := []struct {
+		flags         *appFlags
+		credsReturned bool
+		errorReturned bool
+		message       string
+	}{
+		// success
+		{&appFlags{tls: true}, true, false, ""},
+		{&appFlags{tlsInsecure: true}, true, false, ""},
+		{&appFlags{tlsCAFile: "acctest/x509/certificate.pem"}, true, false, ""},
+		{&appFlags{tlsCAPath: "acctest/x509"}, true, false, ""},
+		// fail
+		{&appFlags{tlsCAFile: "acctest/key.pem"}, false, true, "should fail, acctest/key.pem is not a valid certificate"},
+		{&appFlags{tlsCAFile: "123098.pem"}, false, true, "should fail, 123098.pem does not exist"},
+		{&appFlags{tls: true, tlsInsecure: true}, false, true, "only one tls option should be specified"},
+		{&appFlags{tls: true, tlsCAFile: "acctest/x509/certificate.pem"}, false, true, "only one tls option should be specified"},
+		{&appFlags{tlsInsecure: true, tlsCAFile: "acctest/x509/certificate.pem"}, false, true, "only one tls option should be specified"},
+		{&appFlags{tlsCAFile: "acctest/x509/certificate.pem", tlsCAPath: "acctest/x509"}, false, true, "only one tls option should be specified"},
+	}
+
+	for _, tt := range dataset {
+		// when
+		creds, err := parseCredentials(tt.flags)
+
+		// then
+		if tt.credsReturned {
+			assert.NotNil(t, creds, tt.message)
+		} else {
+			assert.Nil(t, creds, tt.message)
+		}
+
+		if tt.errorReturned {
+			assert.Error(t, err, tt.message)
+		} else {
+			assert.NoError(t, err, tt.message)
+		}
+	}
+}
+
+func Test_countTLSFlags(t *testing.T) {
+	// given
+	dataset := []struct {
+		flags *appFlags
+		count int
+	}{
+		{&appFlags{}, 0},
+		{&appFlags{tls: true}, 1},
+		{&appFlags{tls: true, tlsInsecure: true}, 2},
+		{&appFlags{tls: true, tlsInsecure: true, tlsCAFile: "file"}, 3},
+		{&appFlags{tls: true, tlsInsecure: true, tlsCAFile: "file", tlsCAPath: "path"}, 4},
+	}
+
+	for _, tt := range dataset {
+		// when
+		cnt := countTLSFlags(tt.flags)
+
+		// then
+		assert.Equal(t, tt.count, cnt)
+	}
 }